+++ shared/src/androidMain/kotlin/com/vickikbt/shared/utils/CustomImageLoader.kt
@@ -1,0 +1,32 @@
+package com.vickikbt.shared.utils
+
+import android.content.Context
+import com.seiko.imageloader.ImageLoader
+import com.seiko.imageloader.cache.memory.maxSizePercent
+import com.seiko.imageloader.component.setupDefaultComponents
+import com.seiko.imageloader.option.androidContext
+import okio.Path.Companion.toOkioPath
+
+actual class CustomImageLoader(private val context: Context) {
+
+    actual fun generateImageLoader(): ImageLoader {
+        return ImageLoader {
+            options {
+                androidContext(context)
+            }
+            components {
+                setupDefaultComponents()
+            }
+            interceptor {
+                memoryCacheConfig {
+                    // Set the max size to 25% of the app's available memory.
+                    maxSizePercent(context, 0.25)
+                }
+                diskCacheConfig {
+                    directory(context.cacheDir.resolve("image_cache").toOkioPath())
+                    maxSizeBytes(512L * 1024 * 1024) // 512MB
+                }
+            }
+        }
+    }
+}
