--- shared/src/commonMain/kotlin/com/vickikbt/shared/utils/ImageLoader.kt
@@ -1,5 +1,0 @@
-package com.vickikbt.shared.utils
-
-import com.seiko.imageloader.ImageLoader
-
-expect fun generateImageLoader(): ImageLoader
