--- shared/src/iosMain/kotlin/com/vickikbt/shared/utils/ImageLoader.kt
@@ -1,34 +1,0 @@
-package com.vickikbt.shared.utils
-
-import com.seiko.imageloader.ImageLoader
-import com.seiko.imageloader.component.setupDefaultComponents
-import okio.Path.Companion.toPath
-import platform.Foundation.NSCachesDirectory
-import platform.Foundation.NSSearchPathForDirectoriesInDomains
-import platform.Foundation.NSUserDomainMask
-
-actual fun generateImageLoader(): ImageLoader {
-    return ImageLoader {
-        components {
-            setupDefaultComponents()
-        }
-        interceptor {
-            memoryCacheConfig {
-                maxSizeBytes(32 * 1024 * 1024) // 32MB
-            }
-            diskCacheConfig {
-                directory(getCacheDir().toPath().resolve("image_cache"))
-                maxSizeBytes(512L * 1024 * 1024) // 512MB
-            }
-        }
-    }
-}
-
-private fun getCacheDir(): String {
-    return NSSearchPathForDirectoriesInDomains(
-        NSCachesDirectory,
-        NSUserDomainMask,
-        true,
-    ).first() as String
-}
-
